complete to add  Subject

+++ src/Student/SubjectRegis.java
@@ -4,24 +4,24 @@ import Subject.Subject;
 
 public class SubjectRegis extends Subject {
     private Subject subject;
-    private int grade;
+    private double grade;
     private int year;
     private int term;
 
-    public SubjectRegis(int y,int t,String code, String name, int weight, String[] subjectBase, int grade) {
+    public SubjectRegis(int y,int t,String code, String name, int weight, String[] subjectBase, double grade) {
         super(code, name, weight, subjectBase);
         this.grade = grade;
         term = t;
         year = y;
     }
 
-    public SubjectRegis(int y,int t,String code, String name, int weight, int grade) {
+    public SubjectRegis(int y,int t,String code, String name, int weight, double grade) {
         super(code, name, weight);
         this.grade = grade;
         term = t;
         year = y;
     }
-    public SubjectRegis(int y,int t,Subject subject, int grade) {
+    public SubjectRegis(int y,int t,Subject subject, double grade) {
         super(subject.getCode(), subject.getName(), subject.getWeight());
         this.grade = grade;
         term = t;
@@ -32,7 +32,7 @@ public class SubjectRegis extends Subject {
         return subject;
     }
 
-    public int getGrade() {
+    public double getGrade() {
         return grade;
     }
